requirements.txt
# Core Dependencies
requests>=2.31.0
python-docx>=0.8.11
python-dotenv>=1.0.0
flask>=2.0.0
flask-wtf>=1.2.0
flask-limiter>=4.0.0
wtforms>=3.0.0
limits>=5.6.0
ordered-set>=4.1.0

# Development
pytest>=7.4.0
pytest-mock>=3.11.1
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0